refactor: logout 컴포넌트 이름을 profile로 변경 header에 속해있는 컴포넌트 임으로 header 폴더로 이동 내부적인 상태를 분리하였습니다.
