+++ src/components/organisms/header/profile/hooks/useOutsideClickListener.ts
@@ -1,0 +1,28 @@
+import {
+  BaseSyntheticEvent,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
+
+const useOutsideClickListener = () => {
+  const [isOpen, setIsOpen] = useState(false);
+  const ref = useRef<HTMLDivElement>(null);
+  const handleClickOutside = useCallback(
+    ({ target }: BaseSyntheticEvent | MouseEvent) => {
+      if (ref.current && !ref.current.contains(target)) {
+        setIsOpen(false);
+      }
+    },
+    []
+  );
+  useEffect(() => {
+    window.addEventListener("click", handleClickOutside);
+    return () => {
+      window.removeEventListener("click", handleClickOutside);
+    };
+  }, [handleClickOutside]);
+  return { ref, isOpen, setIsOpen };
+};
+export default useOutsideClickListener;
